power: supply: qbg: Add 32-bit compat ioctl support
When a 32-bit userspace entity calls a IOCTL call on 64-bit Kernel
there can be case where the IOCTL number between user-space and
kernel to be calculated differently because of different type size.

Add support for compat IOCTL which converts the user-space IOCTL
number to kernel supported IOCTL number to fix the above issue.

Also, change the datatype from unsigned long to __u64 for some
struct attributes as the datatype size is calculated differently
in user-space and kernel space.

Change-Id: Idd979b019fe62885038999a5a7d190b253acab83
Signed-off-by: Umang Chheda <quic_uchheda@quicinc.com>
Signed-off-by: Guang Yang <quic_guyang@quicinc.com>