mhi: core: Wait before RDDM transfer complete
Currently MHI stops waiting if the rxvecstatus reg valus is not
zero when downloading rddm img. It's possible that an invalid
non-zero value returns before transfer is completed, in such
scenario, mhi is not able to catch the real completion signal.
Wait until the real transfer completion is coming within the
designated time.

CRs-Fixed: 3430510
Change-Id: I7880f98ba6fce3ca72a839f392165e51670239e3
Signed-off-by: Guisen Yang <quic_guiseny@quicinc.com>