mmc: sdhci-msm: Remove voltage dependency for disabling 3.3v signaling flag
During mmc probe, the vdd-io voltage range was checked to determine
whether the device is emmc or sd card and accordingly 3.3v flag was
disabled/not disabled.

But since the vdd-io range is different for different devices, 3.3v
flag disabling had failed on some devices.

For instance, in case of sa515, the vdd-io range is defined as 2.95v
to 3.6v in dt, but the driver had a check for 2.7 to 3.6v. As a result
the 3.3v signaling flag was not getting disabled.

Therefore, instead of checking voltage range, check for "no-sd" flag
from dt, to determine whether the device is emmc or sd card. The "no-sd"
flag is parsed from dt during early init and accordingly MMC_CAP2_NO_SD
bit is set.

So, add a check for MMC_CAP2_NO_SD bit, instead of the voltage range, to
determine whether the device is emmc or sd card.

Change-Id: I7d270685162403693811ace5ae982496bf046b0e
Signed-off-by: Ram Kumar Dwivedi <quic_rdwivedi@quicinc.com>