serial: msm-geni-serial: Handle xfer_inprogress flag during error exit in UART GSI mode
Ensure the xfer_inprogress flag is cleared during the error exit path in UART GSI mode.
Previously, this flag was not getting reset, causing the TX transfer to remain in progress for subsequent transfers.
This fix addresses the issue by properly handling the xfer_inprogress flag in the error exit path.

Change-Id: Iac5a0c3f4c0b78151418e7611ae073f88c1e8f38
Signed-off-by: Zahir Shabbir Khan <quic_zshabbir@quicinc.com>