serial: msm_geni_serial: Prevent geni register access during suspend
For the BT UART use case, whenever the BT Host closes the port, it calls
vote_clock_off, which internally triggers runtime_suspend, turning off
the clocks. Subsequently, calling TIOCFAULT internally dumps the UART SE
registers. During this time, a race condition is observed where
geni_se_dump_dbg_regs() reads the registers while the UART is suspended
and the clocks are disabled.

It is not expected for ioctl TIOCFAULT to be called by the BT Host after
sending the ioctl call to turn off the clocks. The vote_clock_off function
triggers runtime suspend, which turns off the clocks. When TIOCFAULT is
called, the registers are dumped, causing a NOC error as the clocks are
being turned off simultaneously. To prevent this, a check is added to
avoid dumping the registers when ioctl_count is 0 or the vote clock is
present while calling TIOCFAULT.

Change-Id: Ib278dca585a5bcdf76ceca5597b8354bd0699468
Signed-off-by: Prasanna S <quic_prass@quicinc.com>
Signed-off-by: Visweswara Tanuku <quic_vtanuku@quicinc.com>
Signed-off-by: Panicker Harish <quic_pharish@quicinc.com>