serial: msm_geni_serial: Config/check the QUP registers with active clock votes
For BT UART usecase, whenever BT Host wants to close the port,
first it will call vote_clock_off which will internally call
runtime_suspend and the clocks will be turned off. During this
point of time sometimes race condition is seen where
the serial core is attempting to configure/check the UART
control lines(set_mctrl/get_mctrl), flush the TX buffer
(flush_buffer) or stop the RX sequencer(as part of port close)
due to which NOC error is seen while trying to write/read
to/from the QUP registers.

To prevent this NOC error crash, it needs to be ensured
that the UART control lines are configured/checked, the
TX buffer is flushed, the rx sequencer is started/stopped
only when we have active clock votes.

Earlier, device_pending_suspend() returned true only if the
device was fully suspended but we should also avoid read/write
operations to/from the QUP registers during runtime_resume and
runtime_suspend. It should be ensured that the QUP register access
only happens when the pm runtime state is RPM_ACTIVE using the
pm_runtime_active() API.

Also removed the use of suspend_resume_lock flag and
device_pending_suspend() as it would be redundant now.

Change-Id: I51d120395266f6771c19acf851a3b848adc16544
Signed-off-by: Prasanna S <quic_prass@quicinc.com>
Signed-off-by: Panicker Harish <quic_pharish@quicinc.com>