drivers: sensors: SMI130 hrtimer_forward in handler
Update hrtimer expiry in the timer handler instead of
calling hrtimer_forward from the acc irq handler.
There are some situations when the hrtimer has expired
but the handler is not yet called. The hrtimer handler
(softirq) is preempted by the acc hard irq, and
in the hard irq the hrtimer_forward tries to forward
a timer that is still queued. This change fixes the
issue.

Change-Id: I05f3b7448f7cb32a216ff8801bd749d480ef22ad
Signed-off-by: Saurav Shah <sauravs@qti.qualcomm.com>
(cherry picked from commit aeeb4b7461bdfe62fffdbb96735da9d39bd3f617)