msm_geni_serial: Add ioctl for adding new IPC log in uart
Add ioctl for adding new IPC log buffer in UART driver. If BT
host driver observes a fault in BT transfers this ioctl needs
to be called. On this ioctl call uart driver will change the
logging to new file so that the issue scenario could be preserved
in old UART IPC log files.

Change-Id: I7af38fc7e27b84e8c7cc96015ced34cf498440d3
Signed-off-by: Mukesh Kumar Savaliya <quic_msavaliy@quicinc.com>
Signed-off-by: Yatish Kumar Singh <quic_yatishku@quicinc.com>
Signed-off-by: Panicker Harish <quic_pharish@quicinc.com>