ARM: dts: msm:  adb devices optimize
At boot time, the PD chip is connected to both the CC and VBUS ports,
so the PMIC sends the USB status message to the AP
and then calls the set-role function to set the USB mode.

If the DTS1 configuration specifies bypass-ucsi-enable,
the driver does not set the USB mode and simply returns.

Change-Id: Id48e4d0f3183c113ebbb3f0e70c858902938105b
Signed-off-by: Yi Zhang <zhanyi@qti.qualcomm.com>